pyproject.toml
[project]
name = "flight-blender"
version = "0.1.0"
authors = [{ name = "Hrishikesh Ballal", email = "hrishi@openskies.sh" }]
description = "Flight Blender"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"

keywords = ["Flight Blender"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
]

dependencies = [
    "arrow==1.3.0",
    "celery==5.4.0",
    "cryptography==44.0.1",
    "dacite==1.9.2",
    "dj-database-url==2.1.0",
    "django==5.2.6",
    "django-celery-beat==2.8.1",
    "djangorestframework==3.15.2",
    "geojson==3.2.0",
    "gunicorn==23.0.0",
    "http-message-signatures==0.5.0",
    "implicitdict==2.3.0",
    "jwcrypto==1.5.6",
    "marshmallow==3.25.1",
    "numpy==2.2.1",
    "pandas==2.2.3",
    "pika==1.3.2",
    "psycopg2-binary==2.9.10",
    "pyjwt==2.10.1",
    "pyproj==3.7.2",
    "python-dotenv==1.0.1",
    "python-jose==3.4.0",
    "redis==5.2.1",
    "requests==2.32.5",
    "rtree==1.4.0",
    "shapely==2.0.6",
    "tldextract==5.3.0",
    "uas-standards==3.4.0",
    "wait-for-it==2.3.0",
    "channels==4.3.1", 
    "channels-redis==4.1.5",
]

[dependency-groups]
dev = [
    "bandit>=1.8.3",
    "codespell>=2.4.1",
    "pre-commit>=4.2.0",
    "pylint-django>=2.6.1",
    "pylint-pytest>=1.1.8",
    "pyright>=1.1.398",
    "pyupgrade>=3.19.1",
    "ruff>=0.11.4",
    "python-lsp-server>=1.12.2",
    "pyyaml>=6.0.2",
    "uritemplate>=4.2.0",
    "inflection>=0.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.codespell]
skip = ["**/redis.conf", "uv.lock"]
ignore-words = ".codespellignore"
quiet-level = 2

[tool.pylint]
load-plugins = "pylint_pytest"

[tool.pylint.format]
max-line-length = 150

[tool.pylint."messages control"]
disable = [
    "missing-module-docstring",
    "missing-function-docstring",
    "missing-class-docstring",
    "unused-import",
    "useless-import-alias",
]
